[Fix] Fix i32/i64 issues after rebase

+++ python/tvm/relax/op/nn/nn.py
@@ -618,7 +618,7 @@ def adaptive_avg_pool2d(
             if isinstance(shape, PrimExpr):
                 temp_size.append(shape)
             elif isinstance(shape, int):
-                temp_size.append(tvm.tir.const(shape, "int32"))
+                temp_size.append(tvm.tir.const(shape, "int64"))
             else:
                 raise RuntimeError(
                     f"The input new shape of reshape operator contains unrecognized dimension {shape}"
